Bottom bar icon updation Removed expo icon - added image as icon Code refractoring
