+++ app/(app)/dashboard/_layout.tsx
@@ -1,175 +1,45 @@
 import React from 'react';
 import { Tabs } from 'expo-router';
-import {
-  Image,
-  StyleSheet,
-  Text,
-  TouchableOpacity,
-  View,
-  Platform,
-  Dimensions,
-} from 'react-native';
-import Animated, { useSharedValue, useAnimatedStyle, withSpring } from 'react-native-reanimated';
-import { FontAwesome5, Ionicons } from '@expo/vector-icons';
+import { View, Dimensions, Image } from 'react-native';
+import BottomTab from '@/components/BottomTabBar';
 
 const ROUTES: {
   name: string;
   title: string;
   icon: any;
+  activeIcon: any;
 }[] = [
   {
     name: 'index',
     title: 'Today',
-    // icon: require('@/assets/icons/Today.png'),
-    icon: 'calendar',
+    icon: require('@/assets/icons/today.png'),
+    activeIcon: require('@/assets/icons/today-active.png'),
   },
-  
   {
     name: 'notification',
     title: 'Notify',
-    // icon: require('@/assets/icons/week-month-icon.png'),
-    icon: 'notifications',
+    icon: require('@/assets/icons/bell.png'),
+    activeIcon: require('@/assets/icons/bell-active.png'),
   },
   {
     name: 'add',
     title: 'Add',
-    // icon: require('@/assets/icons/add.png'),
-    icon: 'plus',
+    icon: require('@/assets/icons/add.png'),
+    activeIcon: require('@/assets/icons/add-active.png'),
   },
   {
     name: 'profile',
     title: 'Profile',
-    // icon: require('@/assets/icons/week-month-icon.png'),
-    icon: 'user-alt',
+    icon: require('@/assets/icons/user.png'),
+    activeIcon: require('@/assets/icons/user-active.png'),
   },
   {
     name: 'lends',
     title: 'Lends',
-    // icon: require('@/assets/icons/week-month-icon.png'),
-    icon: 'money-check-alt',
+    icon: require('@/assets/icons/lends.png'),
+    activeIcon: require('@/assets/icons/lends-active.png'),
   },
 ];
-
-function MyTabBar({ state, descriptors, navigation }: any) {
-  return (
-    <View
-      style={{
-        display: 'flex',
-        flexDirection: 'row',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-        elevation: 10,
-        backgroundColor: '#0B0B0F',
-        paddingBottom: Platform.OS === 'ios' ? 10 : 0,
-        // paddingHorizontal: 20,
-        // paddingTop: 10,
-        borderTopColor: '#14141D',
-        borderTopWidth: 1,
-        position: 'static',
-        bottom: 0,
-        // height:90
-      }}>
-      {state.routes.map((route: any, index: number) => {
-        const { options } = descriptors[route.key];
-        const label =
-          options.tabBarLabel !== undefined
-            ? options.tabBarLabel
-            : options.title !== undefined
-              ? options.title
-              : route.name;
-
-        const isFocused = state.index === index;
-
-        const onPress = () => {
-          const event = navigation.emit({
-            type: 'tabPress',
-            target: route.key,
-          });
-
-          if (!isFocused && !event.defaultPrevented) {
-            navigation.navigate(route.name);
-          }
-        };
-
-        const onLongPress = () => {
-          navigation.emit({
-            type: 'tabLongPress',
-            target: route.key,
-          });
-        };
-        const borderColor = useSharedValue('transparent');
-
-        if (isFocused) {
-          borderColor.value = withSpring('#FFCA3A', {
-            duration: 300,
-            dampingRatio: 2,
-          }); // animate when focused
-        } else {
-          borderColor.value = withSpring('transparent'); // animate when focused
-        }
-
-        const animatedStyle = useAnimatedStyle(() => {
-          return {
-            borderTopColor: borderColor.value,
-          };
-        });
-
-        return (
-          <TouchableOpacity
-            accessibilityRole="button"
-            accessibilityState={isFocused ? { selected: true } : {}}
-            accessibilityLabel={options.tabBarAccessibilityLabel}
-            testID={options.tabBarTestID}
-            onPress={onPress}
-            onLongPress={onLongPress}
-            key={route.key}
-            style={[
-              {
-                flex: 1,
-                alignItems: 'center',
-                justifyContent: 'center',
-                marginBottom: 10,
-              },
-            ]}>
-            <Animated.View
-              style={[
-                {
-                  paddingHorizontal: 20,
-                  paddingVertical: 10,
-                  alignItems: 'center',
-                  borderTopWidth: 3,
-                },
-                animatedStyle,
-              ]}>
-              {/* <Image source={options.tabBarIcon} /> */}
-              {['lends','add','profile'].includes(route.name) ? (
-                <FontAwesome5
-                  name={options.tabBarIcon}
-                  size={24}
-                  color={isFocused ? '#FFCA3A' : '#FFF'}
-                />
-              ) : (
-                <Ionicons
-                  name={options.tabBarIcon}
-                  color={isFocused ? '#FFCA3A' : '#FFF'}
-                  size={24}
-                />
-              )}
-            </Animated.View>
-            <Text
-              style={{
-                color: isFocused ? '#FFCA3A' : '#FFF',
-                fontFamily: 'Inter-400',
-                fontSize: 10,
-              }}>
-              {label}
-            </Text>
-          </TouchableOpacity>
-        );
-      })}
-    </View>
-  );
-}
 
 export default function TabLayout() {
   const { width, height } = Dimensions.get('window');
@@ -185,14 +55,15 @@ export default function TabLayout() {
           headerShown: false,
           tabBarShowLabel: false,
         }}
-        tabBar={props => <MyTabBar {...props} />}>
+        tabBar={props => <BottomTab {...props} />}>
         {ROUTES.map(item => (
           <Tabs.Screen
             key={item.name}
             name={item.name}
             options={{
               title: item.title,
-              tabBarIcon: item.icon,
+              tabBarIcon: ({ focused }) =>
+                focused ? <Image source={item.activeIcon} /> : <Image source={item.icon} />,
             }}
           />
         ))}
@@ -200,10 +71,3 @@ export default function TabLayout() {
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  activeColor: {
-    backgroundColor: '#14141D',
-    borderRadius: 10,
-  },
-});
